Complete category, store, brand, age(Get, Post, Delete, Put) and Get …

+++ src/main/java/com/myweb/mamababy/dtos/StoreDTO.java
@@ -1,6 +1,7 @@
 package com.myweb.mamababy.dtos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.Column;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
@@ -27,5 +28,5 @@ public class StoreDTO {
 
     @NotNull(message = "User ID is required")
     @JsonProperty("user_id")
-    private Long userId;
+    private int userId;
 }
